Use a local template file per-repo, instead of a global one
Using a global template file, may have been useful in a mob session where
there are multiple repositories being worked upon. But, the configuration to
use the template is being stored locally. The primary author information is
also going to be stored locally, and it seems best to have the template file
local too.